refactor: Move `JwtProperties` to `config` package for better organization
