+++ src/main/java/com/choogoomoneyna/choogoomoneyna_be/auth/jwt/service/RefreshTokenServiceImpl.java
@@ -1,6 +1,6 @@
 package com.choogoomoneyna.choogoomoneyna_be.auth.jwt.service;
 
-import com.choogoomoneyna.choogoomoneyna_be.auth.jwt.util.JwtProperties;
+import com.choogoomoneyna.choogoomoneyna_be.config.JwtProperties;
 import com.choogoomoneyna.choogoomoneyna_be.auth.jwt.util.JwtTokenProvider;
 import com.choogoomoneyna.choogoomoneyna_be.auth.jwt.mapper.RefreshTokenMapper;
 import com.choogoomoneyna.choogoomoneyna_be.auth.jwt.vo.RefreshTokenVO;
